feat: show resolver on "others" Tab
