+++ components/organisms/RecordsTabBody.tsx
@@ -1,30 +1,12 @@
-import { Button, Modal, Skeleton, RecordItem } from "@ensdomains/thorin";
+import { Skeleton, RecordItem } from "@ensdomains/thorin";
 import { formatDate } from "@/lib/utils/formats";
 import { CoinInfo, DomainData, getCoinNameByType } from "@/lib/domain-page";
-import { useAccount, useEnsName } from "wagmi";
-import { useState } from "react";
-import { EditResolverModalContent } from "./EditResolverModalContent";
-import { useRouter } from "next/router";
-import { excludeKeys } from "@/pages/domains/[name]";
 
 export interface RecordsTabProps {
   domainData: DomainData | null;
 }
 
 export const RecordsTabBody = ({ domainData }: RecordsTabProps) => {
-  const [editResolverModalOpen, setEditResolverModalOpen] = useState(false);
-  const { address } = useAccount();
-
-  const router = useRouter();
-  const { name } = router.query; // Dynamic route parameter
-
-  const { data: authedUserName } = useEnsName({
-    address: address,
-  });
-
-  const showEditButton: boolean =
-    authedUserName === domainData?.owner || address === domainData?.owner;
-
   const millisecondsToSeconds = (millisecodNumber: number): number =>
     millisecodNumber / 1000;
 
@@ -33,17 +15,6 @@ export const RecordsTabBody = ({ domainData }: RecordsTabProps) => {
   const addresses = resolver?.addresses;
   const expiryDate = domainData?.expiryDate;
   const textRecords = domainData?.resolver.texts;
-
-  let filteredRecords: Record<string, string> = {};
-
-  if (domainData && typeof domainData.resolver.texts === "object") {
-    filteredRecords = Object.entries(domainData.resolver.texts)
-      .filter(([key]) => !excludeKeys.includes(key))
-      .reduce((obj: Record<string, string>, [key, value]) => {
-        obj[key] = value as string; // Type assertion to string
-        return obj;
-      }, {});
-  }
 
   return (
     <div className="flex flex-grow flex-col gap-11">
@@ -147,40 +118,6 @@ export const RecordsTabBody = ({ domainData }: RecordsTabProps) => {
 
       <div className="flex flex-col gap-4">
         <Skeleton>
-          <h3 className="text-base font-semibold">Resolver</h3>
-        </Skeleton>
-        <div className="flex flex-col gap-4">
-          <Skeleton>
-            {resolver?.address && (
-              <div className="flex items-center justify-between gap-4 overflow-hidden rounded-md bg-gray-50">
-                <RecordItem
-                  size="large"
-                  keyLabel="resolver"
-                  value={resolver?.address}
-                >
-                  {resolver?.address}
-                </RecordItem>
-
-                {showEditButton && (
-                  <div>
-                    <Button
-                      onClick={() => {
-                        setEditResolverModalOpen(true);
-                      }}
-                      size="small"
-                    >
-                      Edit
-                    </Button>
-                  </div>
-                )}
-              </div>
-            )}
-          </Skeleton>
-        </div>
-      </div>
-
-      <div className="flex flex-col gap-4">
-        <Skeleton>
           <h3 className="text-base font-semibold">Content Hash</h3>
         </Skeleton>
         <div className="flex flex-col gap-4">
@@ -195,17 +132,6 @@ export const RecordsTabBody = ({ domainData }: RecordsTabProps) => {
           </Skeleton>
         </div>
       </div>
-      {resolver?.address && (
-        <Modal open={editResolverModalOpen} onDismiss={() => {}}>
-          <EditResolverModalContent
-            currentResolverAddress={resolver?.address}
-            name={name as string}
-            onCloseModal={() => {
-              setEditResolverModalOpen(false);
-            }}
-          />
-        </Modal>
-      )}
     </div>
   );
 };
